ml_agent/requirements.txt
# Core ML dependencies
torch>=2.4.0
torchvision>=0.19.0
transformers>=4.49.0
diffusers>=0.31.0
tokenizers>=0.20.3
accelerate>=1.1.1

# API and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.9.0.80
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# Utilities
numpy>=1.23.5,<2
psutil>=5.9.0
PyYAML>=6.0.1
tqdm>=4.66.0

# Optional: Hugging Face Hub
huggingface-hub>=0.19.0

# Optional: Flash Attention (for performance)
# flash-attn>=2.7.4  # Uncomment if you have compatible CUDA version
